requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0
scikit-learn>=1.3.0
llama-parse>=0.4.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
spacy>=3.7.0
groq>=0.31.0